code now behaves properly with cooling for gas particles which hit the maximum specified smoothing length (should no longer need to set this arbitrarily high). reads cosmic ray energy from snapshots properly. also restart-writing modified to (hopefully) allow changing multipledomains during restarts - still being tested. couple minor changes to send-receive buffer in hydro to prevent overflows on huge runs.

+++ allvars.c
@@ -358,7 +358,7 @@ struct blackhole_temp_particle_data *BlackholeTempInfo, *BlackholeDataPasserOut,
  * ------------------
  */
 
-int Nexport, Nimport;
+long Nexport, Nimport;
 int BufferFullFlag;
 int NextParticle;
 int NextJ;
